Fixing bugs with authentication & https handling
Fixes #1: user credentials are now passed correctly when making calls to
the server; it does so in the clear so make sure you're on a private
connection or you're using HTTPS to encrypt the connection. Tests
updated

Fixes #2: you are now able to talk to a couchdb instance running over
HTTPS. Tests have been updated. They temporarily allow node to accept
self-signed certificates. This is necessary for my local testing, YMMV
and security concerns may vary