Allow for keys that contain a :
Fixes #3. URLs are allowed to contain colons in the pathname part of the
URL without escaping, but when resolving a URL against a base path, if
you have a colon, you must provide a relative or absolute path (i.e. a
'./' or '/') leading it.

Re-wrote URL resolution to use `path.posix.join` to connect they key and
database strings (to prevent superfluous `/` characters) and only use
url.resolve to attach the server address component to the full path.